Move shared WriteEvent helper
