+++ src/libraries/System.Net.Sockets/src/System/Net/Sockets/NetEventSource.Sockets.cs
@@ -1,7 +1,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Diagnostics.CodeAnalysis;
 using System.Diagnostics.Tracing;
 using System.Net.Sockets;
 using System.Runtime.CompilerServices;
@@ -57,39 +56,5 @@ namespace System.Net
         [NonEvent]
         public static void DumpBuffer(object thisOrContextObject, Memory<byte> buffer, int offset, int count, [CallerMemberName] string? memberName = null) =>
             DumpBuffer(thisOrContextObject, buffer.Span.Slice(offset, count), memberName);
-
-        [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2026:UnrecognizedReflectionPattern",
-                   Justification = EventSourceSuppressMessage)]
-        [NonEvent]
-        private unsafe void WriteEvent(int eventId, string? arg1, string? arg2, int arg3)
-        {
-            arg1 ??= "";
-            arg2 ??= "";
-
-            fixed (char* arg1Ptr = arg1)
-            fixed (char* arg2Ptr = arg2)
-            {
-                const int NumEventDatas = 3;
-                EventData* descrs = stackalloc EventData[NumEventDatas];
-
-                descrs[0] = new EventData
-                {
-                    DataPointer = (IntPtr)(arg1Ptr),
-                    Size = (arg1.Length + 1) * sizeof(char)
-                };
-                descrs[1] = new EventData
-                {
-                    DataPointer = (IntPtr)(arg2Ptr),
-                    Size = (arg2.Length + 1) * sizeof(char)
-                };
-                descrs[2] = new EventData
-                {
-                    DataPointer = (IntPtr)(&arg3),
-                    Size = sizeof(int)
-                };
-
-                WriteEventCore(eventId, NumEventDatas, descrs);
-            }
-        }
     }
 }
